lambda/requirements.txt
discord.py==1.7